requirements.txt
apache-airflow
# Provider specific
apache-airflow-providers-google
pytest-html
PyHamcrest
# For actual testing
pytest
-c "https://raw.githubusercontent.com/apache/airflow/constraints-2.8.3/constraints-3.8.txt"
